+++ app/Http/Services/Decorator/Template2.php
@@ -2,14 +2,17 @@
 namespace App\Http\Services\Decorator;
 
 class Template2 extends HtmlTemplate {
-    protected $_element;
+    protected $_html;
 
-    public function __construct($s) {
-        $this->_element = $s;
-        $this->set("<h2>" . $this->_html . "</h2>");
+    public function __construct() {
+        $this->_html = "<div>__text__</div>";
     }
 
-    public function __call($name, $args) {
-        $this->_element->$name($args[0]);
+    public function set($html) {
+        $this->_html = $html;
+    }
+
+    public function render() {
+        echo $this->_html;
     }
 }
